refactor: consolidated UDL and Nominal files
